backend/requirements.txt
fastapi
uvicorn
pydantic[email]
pydantic-settings
sqlalchemy
pyjwt
passlib[bcrypt]
python-multipart
python-jose[cryptography]
pymongo
dnspython
python-dotenv
google-genai